graph/model: clear small TODOs (gemm_dispatch beta + JSON helper)
graph: gemm_dispatch now honors a beta parameter. Plumbed through the
legacy 23-param dispatch and surfaced via GemmContext::with_beta(). The
four "TODO: migrate to gemm_dispatch with beta=1.0" sites in
executor_attention.cu (Wo + residual) and executor_ffn.cu (w_down +
residual) now route through dispatch instead of inlining gemm()/dequant.
Behaviour unchanged: callers already gate the beta=1 paths on FP16-cache
or dequant_gpu_supported(qtype), which match the dispatch's FP16-cache /
raw-dequant branches.

model: extracted the duplicated minimal JSON parser into
src/model/json_util.{h,cpp}. hf_config_loader now uses the shared header
and drops 250 LOC. The shared parser uses the tokenizer's UTF-16
surrogate-pair handling (strict superset of the previous behaviour, which
collapsed \uXXXX to '?'). safetensors_loader and tokenizer keep their
private copies for now — separate change.

Note: CUDA toolkit not available in this environment; verified
statically. Needs a build + test run on the RTX 5090 host before merge.
