+++ src/model/hf_config_loader.cpp
@@ -1,9 +1,6 @@
 #include "model/hf_config_loader.h"
+#include "model/json_util.h"
 #include "core/logging.h"
-
-#include <fcntl.h>
-#include <sys/stat.h>
-#include <unistd.h>
 
 #include <cstring>
 #include <string>
@@ -11,257 +8,6 @@
 #include <vector>
 
 namespace imp {
-
-// ---- Minimal JSON parser (local copy from safetensors_loader.cpp) ----
-// TODO: extract to a shared json.h header
-
-namespace {
-
-enum class JType { NUL, STRING, NUMBER, ARRAY, OBJECT };
-
-struct JValue {
-    JType type = JType::NUL;
-    std::string str_val;
-    double num_val = 0.0;
-    std::vector<JValue> arr;
-    std::vector<std::pair<std::string, JValue>> obj;
-
-    int64_t as_int() const { return static_cast<int64_t>(num_val); }
-};
-
-class JsonParser {
-public:
-    explicit JsonParser(const char* data, size_t len)
-        : data_(data), len_(len), pos_(0) {}
-
-    JValue parse() {
-        skip_ws();
-        return parse_value();
-    }
-
-    bool ok() const { return !error_; }
-
-private:
-    const char* data_;
-    size_t len_;
-    size_t pos_;
-    bool error_ = false;
-
-    char peek() const {
-        if (pos_ >= len_) return '\0';
-        return data_[pos_];
-    }
-
-    char advance() {
-        if (pos_ >= len_) { error_ = true; return '\0'; }
-        return data_[pos_++];
-    }
-
-    void skip_ws() {
-        while (pos_ < len_ && (data_[pos_] == ' ' || data_[pos_] == '\t' ||
-                                data_[pos_] == '\n' || data_[pos_] == '\r')) {
-            pos_++;
-        }
-    }
-
-    bool expect(char c) {
-        skip_ws();
-        if (peek() == c) { advance(); return true; }
-        error_ = true;
-        return false;
-    }
-
-    JValue parse_value() {
-        skip_ws();
-        if (error_) return {};
-        char c = peek();
-        if (c == '"') return parse_string_value();
-        if (c == '{') return parse_object();
-        if (c == '[') return parse_array();
-        if (c == 't' || c == 'f') return parse_bool();
-        if (c == 'n') return parse_null();
-        if (c == '-' || (c >= '0' && c <= '9')) return parse_number();
-        error_ = true;
-        return {};
-    }
-
-    JValue parse_string_value() {
-        JValue v;
-        v.type = JType::STRING;
-        v.str_val = parse_string_raw();
-        return v;
-    }
-
-    std::string parse_string_raw() {
-        if (!expect('"')) return "";
-        std::string s;
-        while (pos_ < len_) {
-            char c = advance();
-            if (c == '"') return s;
-            if (c == '\\') {
-                if (pos_ >= len_) { error_ = true; return s; }
-                char esc = advance();
-                switch (esc) {
-                    case '"':  s += '"'; break;
-                    case '\\': s += '\\'; break;
-                    case '/':  s += '/'; break;
-                    case 'b':  s += '\b'; break;
-                    case 'f':  s += '\f'; break;
-                    case 'n':  s += '\n'; break;
-                    case 'r':  s += '\r'; break;
-                    case 't':  s += '\t'; break;
-                    case 'u': {
-                        for (int i = 0; i < 4 && pos_ < len_; i++) advance();
-                        s += '?';
-                        break;
-                    }
-                    default: s += esc; break;
-                }
-            } else {
-                s += c;
-            }
-        }
-        error_ = true;
-        return s;
-    }
-
-    JValue parse_number() {
-        JValue v;
-        v.type = JType::NUMBER;
-        size_t start = pos_;
-        if (peek() == '-') advance();
-        while (pos_ < len_ && data_[pos_] >= '0' && data_[pos_] <= '9') advance();
-        if (pos_ < len_ && data_[pos_] == '.') {
-            advance();
-            while (pos_ < len_ && data_[pos_] >= '0' && data_[pos_] <= '9') advance();
-        }
-        if (pos_ < len_ && (data_[pos_] == 'e' || data_[pos_] == 'E')) {
-            advance();
-            if (pos_ < len_ && (data_[pos_] == '+' || data_[pos_] == '-')) advance();
-            while (pos_ < len_ && data_[pos_] >= '0' && data_[pos_] <= '9') advance();
-        }
-        std::string num_str(data_ + start, pos_ - start);
-        v.num_val = std::stod(num_str);
-        return v;
-    }
-
-    JValue parse_object() {
-        JValue v;
-        v.type = JType::OBJECT;
-        if (!expect('{')) return v;
-        skip_ws();
-        if (peek() == '}') { advance(); return v; }
-        while (!error_) {
-            skip_ws();
-            std::string key = parse_string_raw();
-            if (!expect(':')) break;
-            JValue val = parse_value();
-            v.obj.emplace_back(std::move(key), std::move(val));
-            skip_ws();
-            if (peek() == ',') { advance(); continue; }
-            break;
-        }
-        expect('}');
-        return v;
-    }
-
-    JValue parse_array() {
-        JValue v;
-        v.type = JType::ARRAY;
-        if (!expect('[')) return v;
-        skip_ws();
-        if (peek() == ']') { advance(); return v; }
-        while (!error_) {
-            v.arr.push_back(parse_value());
-            skip_ws();
-            if (peek() == ',') { advance(); continue; }
-            break;
-        }
-        expect(']');
-        return v;
-    }
-
-    JValue parse_bool() {
-        JValue v;
-        v.type = JType::NUMBER;
-        if (peek() == 't') {
-            for (int i = 0; i < 4 && pos_ < len_; i++) advance();
-            v.num_val = 1.0;
-        } else {
-            for (int i = 0; i < 5 && pos_ < len_; i++) advance();
-            v.num_val = 0.0;
-        }
-        return v;
-    }
-
-    JValue parse_null() {
-        JValue v;
-        v.type = JType::NUL;
-        for (int i = 0; i < 4 && pos_ < len_; i++) advance();
-        return v;
-    }
-};
-
-const JValue* jobj_find(const JValue& obj, const std::string& key) {
-    for (const auto& kv : obj.obj) {
-        if (kv.first == key) return &kv.second;
-    }
-    return nullptr;
-}
-
-// ---- File I/O helper ----
-
-std::string read_file(const std::string& path) {
-    int fd = open(path.c_str(), O_RDONLY);
-    if (fd < 0) return "";
-    struct stat st;
-    if (fstat(fd, &st) != 0) { close(fd); return ""; }
-    std::string data(st.st_size, '\0');
-    [[maybe_unused]] auto n = ::read(fd, data.data(), st.st_size);
-    close(fd);
-    return data;
-}
-
-// ---- Helper: parse a JSON file into a JValue ----
-
-bool parse_json_file(const std::string& path, JValue& out) {
-    std::string data = read_file(path);
-    if (data.empty()) return false;
-
-    JsonParser parser(data.c_str(), data.size());
-    out = parser.parse();
-    if (!parser.ok() || out.type != JType::OBJECT) {
-        IMP_LOG_WARN("failed to parse JSON: %s", path.c_str());
-        return false;
-    }
-    return true;
-}
-
-// ---- Helper: extract numeric field from JSON object ----
-
-template<typename T>
-bool jobj_get_int(const JValue& obj, const std::string& key, T& out) {
-    const JValue* v = jobj_find(obj, key);
-    if (!v || v->type != JType::NUMBER) return false;
-    out = static_cast<T>(v->num_val);
-    return true;
-}
-
-bool jobj_get_float(const JValue& obj, const std::string& key, float& out) {
-    const JValue* v = jobj_find(obj, key);
-    if (!v || v->type != JType::NUMBER) return false;
-    out = static_cast<float>(v->num_val);
-    return true;
-}
-
-bool jobj_get_string(const JValue& obj, const std::string& key, std::string& out) {
-    const JValue* v = jobj_find(obj, key);
-    if (!v || v->type != JType::STRING) return false;
-    out = v->str_val;
-    return true;
-}
-
-} // anonymous namespace
 
 // ---- Architecture mapping ----
